requirements.txt
stripe>=7.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
psutil>=5.9.0
